IOS-00088 Workout iOS 26 Hybrid Architecture IOS-00088G Fix
